# backend-service/src/main/resources/static/docs/openapi.json
{
  "components": {
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
    "/version": {
      "get": {
        "description": "Returns the current API version",
        "operationId": "getVersion",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "note"
        ]
      }
    },
    "/note/{id}": {
      "get": {
        "description": "Retrives a note by id.",
        "operationId": "getNote",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The UUID of the note to retrieve.",
            "required": true
          }
        ],
        "tags": [
          "note"
        ]
      },
      "delete": {
        "description": "Deletes a note by id.",
        "operationId": "deleteNote",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The UUID of the note to delete.",
            "required": true
          }
        ],
        "tags": [
          "note"
        ]
      }
    },
    "/note": {
      "post": {
        "requestBody": {
          "description": "The note to be created.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Note"
              }
            }
          },
          "required": true
        },
        "description": "Adds a new Note.",
        "operationId": "addNote",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          }
        },
        "tags": [
          "note"
        ]
      }
    }
  },
  "info": {
    "contact": {
      "name": "Steve Mitchell",
      "url": "https://smitchell.github.io",
      "email": "stevecmitchell@gmail.com"
    },
    "description": "The NoteController API was created to work with Notes via REST",
    "title": "Notes REST API",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "note",
      "description": "Note API"
    }
  ]
}